+++ resources/views/livewire/admin/user/index.blade.php
@@ -22,7 +22,9 @@
                             <h5 class="card-title fw-semibold mb-4">Project Listing</h5>
                         </div> --}}
                         <div class="ms-auto">
-                            <a href="#" data-bs-toggle="modal" data-bs-target="#addUserModal" class="btn btn-sm btn-primary text-white"><i class="fas fa-plus-square pr-4"></i>&nbsp;&nbsp; Add User</a>
+                            @if ($admin->isAdmin)
+                                <a href="#" data-bs-toggle="modal" data-bs-target="#addUserModal" class="btn btn-sm btn-primary text-white"><i class="fas fa-plus-square pr-4"></i>&nbsp;&nbsp; Add User</a>
+                            @endif
                         </div>
                     </div>
                     @if (session('message'))
@@ -45,7 +47,9 @@
                                     <th class="text-secondary text-xs font-weight-semibold opacity-7 col-4">Name</th>
                                     <th class="text-secondary text-xs font-weight-semibold opacity-7 col-4">Email Address</th>
                                     <th class="text-secondary text-xs font-weight-semibold opacity-7 col-1 text-center">Status</th>
-                                    <th class="text-secondary text-xs font-weight-semibold opacity-7 col-1">Action</th>
+                                    @if ($admin->isAdmin)
+                                        <th class="text-secondary text-xs font-weight-semibold opacity-7 col-1">Action</th>
+                                    @endif
                                 </tr>
                             </thead>
                             <tbody>
@@ -66,14 +70,14 @@
                                                 <i class="fas fa-ban text-danger"></i>
                                             @endif
                                         </td>
-                                        <td class="">
-
-
-                                            <div class="btn-group" role="group" aria-label="Basic example">
-                                                <a href="#" wire:click="editUser({{$user->id}})" data-bs-toggle="modal" data-bs-target="#editUserModal" class="btn btn-sm btn-warning text-white"><i class="fas fa-pen"></i></a>
-                                                <a href="#" wire:click="deleteUser({{$user->id}})" data-bs-toggle="modal" data-bs-target="#deleteUserModal" class="btn btn-sm btn-danger text-white"><i class="fas fa-trash-alt"></i></a>
-                                            </div>
-                                        </td>
+                                        @if ($admin->isAdmin)
+                                            <td class="">
+                                                <div class="btn-group" role="group" aria-label="Basic example">
+                                                    <a href="#" wire:click="editUser({{$user->id}})" data-bs-toggle="modal" data-bs-target="#editUserModal" class="btn btn-sm btn-warning text-white"><i class="fas fa-pen"></i></a>
+                                                    <a href="#" wire:click="deleteUser({{$user->id}})" data-bs-toggle="modal" data-bs-target="#deleteUserModal" class="btn btn-sm btn-danger text-white"><i class="fas fa-trash-alt"></i></a>
+                                                </div>
+                                            </td>
+                                        @endif
                                     </tr>
                                 @empty
                                     <tr>
